Upgrades solution to VS 2017
This commit upgrades Celeris to VS studio 2017. This is an important commit because it will make further developments easier. But it is a transitional commit and therefore the software version is set to X. In the next steps I will boost performance of Celeris and remove some performance bottlenecks, then I will release a new version.