Removes an interaction limitation.
In previous versions to interact with the model (add water, etc.) the camera had to be fairly close to the terrain. This commit removes this limitation.
Furthermore, a previous commit promised some optimizations in the software. Those optimizations are postponed to a later time, and are not in progress at this point.